pyproject.toml
[tool.poetry]
name = "chdlm-mini-dashboard"
version = "0.1.0"
description = ""
authors = ["Yannick Drolet <ydrolet@gmail.com>"]
readme = "README.adoc"

[tool.poetry.dependencies]
python = "^3.12"
nicegui = "^1.4.35"
google-api-python-client = "^2.141.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
pydantic-settings = "^2.4.0"
pendulum = "^3.0.0"
pydantic-extra-types = "^2.9.0"
mjml-python = "^1.3.3"
jinja2 = "^3.1.4"
email-validator = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
chdlm_mini_dashboard = "chdlm_mini_dashboard.main:main"
